fix(cli): the claim keys on the spelling the CLI reads a deployment by
`normalizeApiUrl` strips a query, a fragment, empty path segments and a
trailing path slash, so the CLI already reads six spellings of one host as
one deployment. The claim keyed on `new URL(apiUrl).href`, which keeps four
of them apart. Where the two disagree is exactly where it hurts: the
normalization runs in `parsePieceOptions` and in `cf wish`, both in the
command layer, and nothing under `lib/` calls it — so a caller that holds
its own `SpaceConfig` and reaches `loadPieces` through the export entries
gets no normalization at all, and a shell that reconnected to
`https://host/base` after opening on `https://host/base/` was refused by
the guard meant to protect it.

One normalization, not two. `normalizeApiUrl` moves to `lib/api-url.ts`,
where both layers can reach it without `lib/` importing from `commands/`,
and it gains the doc comment it never had. Its three callers name the new
module directly rather than a re-export, and its own test moves with the
import but keeps its name and its place, so nothing about it is renamed.
The decline still runs first: a URL no connection can be opened over is
declined before `normalizeApiUrl`, which throws on exactly those.

Two cases change meaning with the key. The refusal now names each
deployment in its normalized spelling, so the case reading the message
reads it without the trailing slash `href` used to add. And a new case
pins what the finding named: the spellings the CLI reads as one deployment
claim one deployment.
