+++ packages/cli/test/piece.test.ts
@@ -29,7 +29,6 @@ import {
   getCellValueFromCommand,
   localPatternEntry,
   mergePiecePath,
-  normalizeApiUrl,
   parseLink,
   parsePieceOptions,
   parseSpaceOptions,
@@ -40,6 +39,7 @@ import {
   setCellValueFromCommand,
   setPieceSourceFromCommand,
 } from "../commands/piece.ts";
+import { normalizeApiUrl } from "../lib/api-url.ts";
 import {
   CellSelectionError,
   parseCellSelectionOptions,
